requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mutagen==1.47.0
mysql-connector-python==8.4.0
numpy==1.23.5
openai==1.31.0
packaging==23.2
pillow==10.4.0
pip-search==0.0.12
pycparser==2.22
pycryptodomex==3.20.0
Pygments==2.18.0
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.8.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2024.4.9